+++ packages/contracts/contracts/libs/OperationUtils.sol
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 pragma solidity ^0.8.17;
 import {Groth16} from "../libs/Groth16.sol";
-import {Utils} from "../libs/Utils.sol";
 import "../libs/Types.sol";
 
 // Helpers for extracting data / formatting operations